feat(gdextension): complete metadata-backed phase 0 contracts
- Parse 64-bit enum values, global constants, and builtin layout metadata as shared facts.
- Route engine property access, static-load, and method-bind codegen through metadata-backed resolvers.
- Preserve frontend opaque-route binding provenance without leaking it into CFG public payloads.
- Add focused backend, frontend, gdextension, LIR, and scope tests for edge and failure cases.
- Update phase 0 and module docs with the enforced semantic boundaries.

+++ src/main/java/gd/script/gdcc/gdextension/ExtensionEnumValue.java
@@ -1,6 +1,8 @@
 package gd.script.gdcc.gdextension;
 
+/// Godot exposes extension API enum values as 64-bit Variant::INT/CoreConstants values.
 public record ExtensionEnumValue(
         String name,
-        int value
-) { }
+        long value
+) {
+}
